New reference rect style.
