+++ src/core/src/render/renderSpan.js
@@ -1,4 +1,5 @@
 import { renderKnob } from './renderKnob'
+import { hex2rgba } from '../utils/color'
 
 /**
  * Create a Span element.
@@ -42,33 +43,11 @@ function createRect (r, color, readOnly) {
         const rgba = hex2rgba(color, 0.4)
 
         return $('<div class="anno-span__area"/>').css({
-            top                   : r.y + 'px',
-            left                  : r.x + 'px',
-            width                 : r.width + 'px',
-            height                : r.height + 'px',
-            backgroundBottomColor : rgba
+            top             : r.y + 'px',
+            left            : r.x + 'px',
+            width           : r.width + 'px',
+            height          : r.height + 'px',
+            backgroundColor : rgba
         })
     }
-}
-
-/**
- * Change color definition style from hex to rgba.
- */
-function hex2rgba (hex, alpha = 1) {
-
-    // long version
-    let r = hex.match(/^#([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})$/i)
-    let c = null
-    if (r) {
-        c = r.slice(1, 4).map(function (x) { return parseInt(x, 16) })
-    }
-    // short version
-    r = hex.match(/^#([0-9a-f])([0-9a-f])([0-9a-f])$/i)
-    if (r) {
-        c = r.slice(1, 4).map(function (x) { return 0x11 * parseInt(x, 16) })
-    }
-    if (!c) {
-        return hex
-    }
-    return `rgba(${c[0]}, ${c[1]}, ${c[2]}, ${alpha})`
 }
